flask/requirements.txt
azure-cognitiveservices-speech==1.7.0
astroid==2.3.2
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
firebase-admin==3.1.0
Flask==1.1.1
Flask-Cors==3.0.8
google-api-core==1.14.3
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.5.0
google-cloud-storage==1.21.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
grpcio==1.24.3
httplib2==0.14.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.2
numpy==1.15.4
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pylint==2.4.3
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.12.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.3
Werkzeug==0.15.5
wrapt==1.11.2
